oracle migration: Rearrange code
- Refactoring commit, no functional changes
- Reformat and move code around to have related code together
