+++ backend/hitas/oracle_migration/oracle_schema/company.py
@@ -79,13 +79,7 @@ company_addresses = Table(
     "HITYHTIONOSOITE",
     metadata_obj,
     Column("KG_YOTUNNUS", Integer, key="id", primary_key=True),
-    Column(
-        "KG_YTUNNUS",
-        Integer,
-        ForeignKey("HITYHTIO.id"),
-        key="company_id",
-        nullable=False,
-    ),
+    Column("KG_YTUNNUS", Integer, ForeignKey("HITYHTIO.id"), key="company_id", nullable=False),
     Column("C_KATUOS", String(50), nullable=False),
     Column("C_MUUTTAJA", String(10), nullable=False),
     Column("D_MUUTETTU", Date, nullable=False),
@@ -95,13 +89,7 @@ real_estates = Table(
     "HITKIINTEISTO",
     metadata_obj,
     Column("KG_KTUNNUS", Integer, key="id", primary_key=True),
-    Column(
-        "KG_YTUNNUS",
-        Integer,
-        ForeignKey("HITYHTIO.id"),
-        key="company_id",
-        nullable=False,
-    ),
+    Column("KG_YTUNNUS", Integer, ForeignKey("HITYHTIO.id"), key="company_id", nullable=False),
     Column("C_KIINTTUN", HitasAnonymizedPropertyIdentifier(14), key="property_identifier", nullable=False),
     Column("C_KUNTA", String(3), nullable=False),
     Column("C_KOSAKOODI", String(16), nullable=False),  # Always 'KAUPOSA'
@@ -117,14 +105,7 @@ apartments = Table(
     "HITHUONE",
     metadata_obj,
     Column("KG_HTUNNUS", Integer, key="id", primary_key=True),
-    Column(
-        "KG_YTUNNUS",
-        Integer,
-        ForeignKey("HITYHTIO.id"),
-        key="company_id",
-        nullable=False,
-        index=True,
-    ),
+    Column("KG_YTUNNUS", Integer, ForeignKey("HITYHTIO.id"), key="company_id", nullable=False, index=True),
     Column("C_OMNIMI1", HitasAnonymizedName(50), index=True),
     Column("C_SOTU1", HitasAnonymizedSSN(11)),
     Column("C_OMNIMI2", HitasAnonymizedName(50), index=True),
